+++ src/cmd.rs
@@ -7,14 +7,14 @@ use ash;
 /// barriers to be passed to `vkCmdPipelineBarrier`.
 /// `command_buffer` is passed unmodified to `vkCmdPipelineBarrier`.
 pub fn pipeline_barrier(
-	device: ash::vk::cmds::DeviceFnV1_0,
+	device: ash::vk::DeviceFnV1_0,
 	command_buffer: ash::vk::CommandBuffer,
 	global_barrier: Option<GlobalBarrier>,
 	buffer_barriers: &[BufferBarrier],
 	image_barriers: &[ImageBarrier],
 ) {
-	let mut src_stage_mask = ash::vk::PIPELINE_STAGE_TOP_OF_PIPE_BIT;
-	let mut dst_stage_mask = ash::vk::PIPELINE_STAGE_BOTTOM_OF_PIPE_BIT;
+	let mut src_stage_mask = ash::vk::PipelineStageFlags::TOP_OF_PIPE;
+	let mut dst_stage_mask = ash::vk::PipelineStageFlags::BOTTOM_OF_PIPE;
 
 	let mut vk_memory_barriers: Vec<ash::vk::MemoryBarrier> = Vec::with_capacity(1);
 	let mut vk_buffer_barriers: Vec<ash::vk::BufferMemoryBarrier> =
@@ -66,12 +66,12 @@ pub fn pipeline_barrier(
 /// Sets an event when the accesses defined by `previous_accesses` are completed.
 /// `command_buffer` and `event` are passed unmodified to `vkCmdSetEvent`.
 pub fn set_event(
-	device: ash::vk::cmds::DeviceFnV1_0,
+	device: ash::vk::DeviceFnV1_0,
 	command_buffer: ash::vk::CommandBuffer,
 	event: ash::vk::Event,
 	previous_accesses: &[AccessType],
 ) {
-	let mut stage_mask = ash::vk::PIPELINE_STAGE_TOP_OF_PIPE_BIT;
+	let mut stage_mask = ash::vk::PipelineStageFlags::TOP_OF_PIPE;
 	for previous_access in previous_accesses {
 		let previous_info = get_access_info(previous_access);
 		stage_mask |= previous_info.stage_mask;
@@ -86,12 +86,12 @@ pub fn set_event(
 /// Resets an event when the accesses defined by `previous_accesses` are completed.
 /// `command_buffer` and `event` are passed unmodified to `vkCmdResetEvent`.
 pub fn reset_event(
-	device: ash::vk::cmds::DeviceFnV1_0,
+	device: ash::vk::DeviceFnV1_0,
 	command_buffer: ash::vk::CommandBuffer,
 	event: ash::vk::Event,
 	previous_accesses: &[AccessType],
 ) {
-	let mut stage_mask = ash::vk::PIPELINE_STAGE_TOP_OF_PIPE_BIT;
+	let mut stage_mask = ash::vk::PipelineStageFlags::TOP_OF_PIPE;
 	for previous_access in previous_accesses {
 		let previous_info = get_access_info(previous_access);
 		stage_mask |= previous_info.stage_mask;
@@ -109,15 +109,15 @@ pub fn reset_event(
 ///
 /// `commandBuffer` and `events` are passed unmodified to `vkCmdWaitEvents`.
 pub fn wait_events(
-	device: ash::vk::cmds::DeviceFnV1_0,
+	device: ash::vk::DeviceFnV1_0,
 	command_buffer: ash::vk::CommandBuffer,
 	events: &[ash::vk::Event],
 	global_barrier: Option<GlobalBarrier>,
 	buffer_barriers: &[BufferBarrier],
 	image_barriers: &[ImageBarrier],
 ) {
-	let mut src_stage_mask = ash::vk::PIPELINE_STAGE_TOP_OF_PIPE_BIT;
-	let mut dst_stage_mask = ash::vk::PIPELINE_STAGE_BOTTOM_OF_PIPE_BIT;
+	let mut src_stage_mask = ash::vk::PipelineStageFlags::TOP_OF_PIPE;
+	let mut dst_stage_mask = ash::vk::PipelineStageFlags::BOTTOM_OF_PIPE;
 
 	let mut vk_memory_barriers: Vec<ash::vk::MemoryBarrier> = Vec::with_capacity(1);
 	let mut vk_buffer_barriers: Vec<ash::vk::BufferMemoryBarrier> =