+++ tests/global.rs
@@ -13,10 +13,10 @@ fn compute_write_storage_compute_read_storage() {
 
 	let (src_mask, dst_mask, barrier) = vk_sync::get_memory_barrier(&global_barrier);
 
-	assert_eq!(src_mask, ash::vk::PIPELINE_STAGE_COMPUTE_SHADER_BIT);
-	assert_eq!(dst_mask, ash::vk::PIPELINE_STAGE_COMPUTE_SHADER_BIT);
-	assert_eq!(barrier.src_access_mask, ash::vk::ACCESS_SHADER_WRITE_BIT);
-	assert_eq!(barrier.dst_access_mask, ash::vk::ACCESS_SHADER_READ_BIT);
+	assert_eq!(src_mask, ash::vk::PipelineStageFlags::COMPUTE_SHADER);
+	assert_eq!(dst_mask, ash::vk::PipelineStageFlags::COMPUTE_SHADER);
+	assert_eq!(barrier.src_access_mask, ash::vk::AccessFlags::SHADER_WRITE);
+	assert_eq!(barrier.dst_access_mask, ash::vk::AccessFlags::SHADER_READ);
 }
 
 #[test]
@@ -29,10 +29,10 @@ fn compute_read_storage_compute_write_storage() {
 
 	let (src_mask, dst_mask, barrier) = vk_sync::get_memory_barrier(&global_barrier);
 
-	assert_eq!(src_mask, ash::vk::PIPELINE_STAGE_COMPUTE_SHADER_BIT);
-	assert_eq!(dst_mask, ash::vk::PIPELINE_STAGE_COMPUTE_SHADER_BIT);
-	assert_eq!(barrier.src_access_mask, ash::vk::ACCESS_SHADER_WRITE_BIT);
-	assert_eq!(barrier.dst_access_mask, ash::vk::ACCESS_SHADER_READ_BIT);
+	assert_eq!(src_mask, ash::vk::PipelineStageFlags::COMPUTE_SHADER);
+	assert_eq!(dst_mask, ash::vk::PipelineStageFlags::COMPUTE_SHADER);
+	assert_eq!(barrier.src_access_mask, ash::vk::AccessFlags::SHADER_WRITE);
+	assert_eq!(barrier.dst_access_mask, ash::vk::AccessFlags::SHADER_READ);
 }
 
 #[test]
@@ -45,10 +45,10 @@ fn compute_write_storage_graphics_read_index() {
 
 	let (src_mask, dst_mask, barrier) = vk_sync::get_memory_barrier(&global_barrier);
 
-	assert_eq!(src_mask, ash::vk::PIPELINE_STAGE_COMPUTE_SHADER_BIT);
-	assert_eq!(dst_mask, ash::vk::PIPELINE_STAGE_VERTEX_INPUT_BIT);
-	assert_eq!(barrier.src_access_mask, ash::vk::ACCESS_SHADER_WRITE_BIT);
-	assert_eq!(barrier.dst_access_mask, ash::vk::ACCESS_INDEX_READ_BIT);
+	assert_eq!(src_mask, ash::vk::PipelineStageFlags::COMPUTE_SHADER);
+	assert_eq!(dst_mask, ash::vk::PipelineStageFlags::VERTEX_INPUT);
+	assert_eq!(barrier.src_access_mask, ash::vk::AccessFlags::SHADER_WRITE);
+	assert_eq!(barrier.dst_access_mask, ash::vk::AccessFlags::INDEX_READ);
 }
 
 #[test]
@@ -61,12 +61,12 @@ fn compute_write_storage_graphics_read_indirect() {
 
 	let (src_mask, dst_mask, barrier) = vk_sync::get_memory_barrier(&global_barrier);
 
-	assert_eq!(src_mask, ash::vk::PIPELINE_STAGE_COMPUTE_SHADER_BIT);
-	assert_eq!(dst_mask, ash::vk::PIPELINE_STAGE_DRAW_INDIRECT_BIT);
-	assert_eq!(barrier.src_access_mask, ash::vk::ACCESS_SHADER_WRITE_BIT);
+	assert_eq!(src_mask, ash::vk::PipelineStageFlags::COMPUTE_SHADER);
+	assert_eq!(dst_mask, ash::vk::PipelineStageFlags::DRAW_INDIRECT);
+	assert_eq!(barrier.src_access_mask, ash::vk::AccessFlags::SHADER_WRITE);
 	assert_eq!(
 		barrier.dst_access_mask,
-		ash::vk::ACCESS_INDIRECT_COMMAND_READ_BIT
+		ash::vk::AccessFlags::INDIRECT_COMMAND_READ
 	);
 }
 
@@ -80,8 +80,8 @@ fn nothing_transfer_read() {
 
 	let (src_mask, dst_mask, barrier) = vk_sync::get_memory_barrier(&global_barrier);
 
-	assert_eq!(src_mask, ash::vk::PIPELINE_STAGE_TOP_OF_PIPE_BIT);
-	assert_eq!(dst_mask, ash::vk::PIPELINE_STAGE_TRANSFER_BIT);
+	assert_eq!(src_mask, ash::vk::PipelineStageFlags::TOP_OF_PIPE);
+	assert_eq!(dst_mask, ash::vk::PipelineStageFlags::TRANSFER);
 	assert_eq!(barrier.src_access_mask, ash::vk::AccessFlags::empty());
 	assert_eq!(barrier.dst_access_mask, ash::vk::AccessFlags::empty());
 }
@@ -96,12 +96,12 @@ fn transfer_write_graphics_read_vertex() {
 
 	let (src_mask, dst_mask, barrier) = vk_sync::get_memory_barrier(&global_barrier);
 
-	assert_eq!(src_mask, ash::vk::PIPELINE_STAGE_TRANSFER_BIT);
-	assert_eq!(dst_mask, ash::vk::PIPELINE_STAGE_VERTEX_INPUT_BIT);
-	assert_eq!(barrier.src_access_mask, ash::vk::ACCESS_TRANSFER_WRITE_BIT);
+	assert_eq!(src_mask, ash::vk::PipelineStageFlags::TRANSFER);
+	assert_eq!(dst_mask, ash::vk::PipelineStageFlags::VERTEX_INPUT);
+	assert_eq!(barrier.src_access_mask, ash::vk::AccessFlags::TRANSFER_WRITE);
 	assert_eq!(
 		barrier.dst_access_mask,
-		ash::vk::ACCESS_VERTEX_ATTRIBUTE_READ_BIT
+		ash::vk::AccessFlags::VERTEX_ATTRIBUTE_READ
 	);
 }
 
@@ -115,15 +115,15 @@ fn full_pipeline_barrier() {
 
 	let (src_mask, dst_mask, barrier) = vk_sync::get_memory_barrier(&global_barrier);
 
-	assert_eq!(src_mask, ash::vk::PIPELINE_STAGE_ALL_COMMANDS_BIT);
-	assert_eq!(dst_mask, ash::vk::PIPELINE_STAGE_ALL_COMMANDS_BIT);
+	assert_eq!(src_mask, ash::vk::PipelineStageFlags::ALL_COMMANDS);
+	assert_eq!(dst_mask, ash::vk::PipelineStageFlags::ALL_COMMANDS);
 	assert_eq!(
 		barrier.src_access_mask,
-		ash::vk::ACCESS_MEMORY_READ_BIT | ash::vk::ACCESS_MEMORY_WRITE_BIT
+		ash::vk::AccessFlags::MEMORY_READ | ash::vk::AccessFlags::MEMORY_WRITE
 	);
 	assert_eq!(
 		barrier.dst_access_mask,
-		ash::vk::ACCESS_MEMORY_READ_BIT | ash::vk::ACCESS_MEMORY_WRITE_BIT
+		ash::vk::AccessFlags::MEMORY_READ | ash::vk::AccessFlags::MEMORY_WRITE
 	);
 }
 
@@ -140,15 +140,15 @@ fn compute_write_storage_graphics_read_index_compute_read_uniform() {
 
 	let (src_mask, dst_mask, barrier) = vk_sync::get_memory_barrier(&global_barrier);
 
-	assert_eq!(src_mask, ash::vk::PIPELINE_STAGE_COMPUTE_SHADER_BIT);
+	assert_eq!(src_mask, ash::vk::PipelineStageFlags::COMPUTE_SHADER);
 	assert_eq!(
 		dst_mask,
-		ash::vk::PIPELINE_STAGE_VERTEX_INPUT_BIT | ash::vk::PIPELINE_STAGE_COMPUTE_SHADER_BIT
+		ash::vk::PipelineStageFlags::VERTEX_INPUT | ash::vk::PipelineStageFlags::COMPUTE_SHADER
 	);
-	assert_eq!(barrier.src_access_mask, ash::vk::ACCESS_SHADER_WRITE_BIT);
+	assert_eq!(barrier.src_access_mask, ash::vk::AccessFlags::SHADER_WRITE);
 	assert_eq!(
 		barrier.dst_access_mask,
-		ash::vk::ACCESS_INDEX_READ_BIT | ash::vk::ACCESS_UNIFORM_READ_BIT
+		ash::vk::AccessFlags::INDEX_READ | ash::vk::AccessFlags::UNIFORM_READ
 	);
 }
 
@@ -165,14 +165,14 @@ fn compute_write_texel_graphics_read_indirect_fragment_read_uniform() {
 
 	let (src_mask, dst_mask, barrier) = vk_sync::get_memory_barrier(&global_barrier);
 
-	assert_eq!(src_mask, ash::vk::PIPELINE_STAGE_COMPUTE_SHADER_BIT);
+	assert_eq!(src_mask, ash::vk::PipelineStageFlags::COMPUTE_SHADER);
 	assert_eq!(
 		dst_mask,
-		ash::vk::PIPELINE_STAGE_DRAW_INDIRECT_BIT | ash::vk::PIPELINE_STAGE_FRAGMENT_SHADER_BIT
+		ash::vk::PipelineStageFlags::DRAW_INDIRECT | ash::vk::PipelineStageFlags::FRAGMENT_SHADER
 	);
-	assert_eq!(barrier.src_access_mask, ash::vk::ACCESS_SHADER_WRITE_BIT);
+	assert_eq!(barrier.src_access_mask, ash::vk::AccessFlags::SHADER_WRITE);
 	assert_eq!(
 		barrier.dst_access_mask,
-		ash::vk::ACCESS_INDIRECT_COMMAND_READ_BIT | ash::vk::ACCESS_UNIFORM_READ_BIT
+		ash::vk::AccessFlags::INDIRECT_COMMAND_READ | ash::vk::AccessFlags::UNIFORM_READ
 	);
 }