+++ CHANGES.md
@@ -1,5 +1,13 @@
 # Changes
 
+## 0.1.1 (2018-11-15)
+
+* Updated to ash 0.25 (Vulkan 1.1)
+* Added support for NVX generated commands
+* Added support for read-only depth/stencil + writeable depth/stencil
+* Added Copy and Default traits to AccessType and ImageLayout
+* Added Debug, Default, and Clone traits to GlobalBarrier, BufferBarrier, and ImageBarrier
+
 ## 0.1.0 (2018-08-26)
 
 * First release